Updating svg component
