+++ src/components/ui/LinkButton/LinkButton.jsx
@@ -4,10 +4,10 @@ import './LinkButton.css'
 import { Svg } from '../Svg/Svg'
 /* eslint-enable no-unused-vars */
 
-export function LinkButton ({ text, url, svgName, svgSize = 15, svgColors = ['#f3f4f6'] }) {
+export function LinkButton ({ text, url, svgName, svgSize = 15 }) {
     return (
         <a href={url} className='link-button' target='_blank' rel='noopener noreferrer'>
-            <Svg name={svgName} size={svgSize} colors={svgColors} />
+            <Svg name={svgName} size={svgSize} />
             <span>{text}</span>
         </a>
     )
