+++ src/components/ui/Subtitle/Subtitle.jsx
@@ -4,10 +4,10 @@ import './Subtitle.css'
 import { Svg } from '../Svg/Svg'
 /* eslint-enable no-unused-vars */
 
-export function Subtitle ({ text, svgName, svgSize = 30, svgColors = ['#f3f4f6'] }) {
+export function Subtitle ({ text, svgName, svgSize = 30 }) {
     return (
         <div className='subtitle'>
-            <Svg name={svgName} size={svgSize} colors={svgColors} />
+            <Svg name={svgName} size={svgSize} />
             <h2>{text}</h2>
         </div>
     )
