+++ src/components/ui/Svg/svgMap.js
@@ -1,0 +1,47 @@
+// svgMap = [[d1, ... , dr], viewbox]
+export const svgMap = {
+    lightMode: [
+        ['M480-360q50 0 85-35t35-85q0-50-35-85t-85-35q-50 0-85 35t-35 85q0 50 35 85t85 35Zm0 80q-83 0-141.5-58.5T280-480q0-83 58.5-141.5T480-680q83 0 141.5 58.5T680-480q0 83-58.5 141.5T480-280ZM200-440H40v-80h160v80Zm720 0H760v-80h160v80ZM440-760v-160h80v160h-80Zm0 720v-160h80v160h-80ZM256-650l-101-97 57-59 96 100-52 56Zm492 496-97-101 53-55 101 97-57 59Zm-98-550 97-101 59 57-100 96-56-52ZM154-212l101-97 55 53-97 101-59-57Zm326-268Z'],
+        '0 -960 960 960'
+    ],
+    darkMode: [
+        ['M480-120q-150 0-255-105T120-480q0-150 105-255t255-105q14 0 27.5 1t26.5 3q-41 29-65.5 75.5T444-660q0 90 63 153t153 63q55 0 101-24.5t75-65.5q2 13 3 26.5t1 27.5q0 150-105 255T480-120Zm0-80q88 0 158-48.5T740-375q-20 5-40 8t-40 3q-123 0-209.5-86.5T364-660q0-20 3-40t8-40q-78 32-126.5 102T200-480q0 116 82 198t198 82Zm-10-270Z'],
+        '0 -960 960 960'
+    ],
+    systemMode: [
+        ['M240-120v-80l40-40H160q-33 0-56.5-23.5T80-320v-440q0-33 23.5-56.5T160-840h640q33 0 56.5 23.5T880-760v440q0 33-23.5 56.5T800-240H680l40 40v80H240Zm-80-200h640v-440H160v440Zm0 0v-440 440Z'],
+        '0 -960 960 960'
+    ],
+    experience: [
+        ['M160-200v-440 440-15 15Zm0 80q-33 0-56.5-23.5T80-200v-440q0-33 23.5-56.5T160-720h160v-80q0-33 23.5-56.5T400-880h160q33 0 56.5 23.5T640-800v80h160q33 0 56.5 23.5T880-640v171q-18-13-38-22.5T800-508v-132H160v440h283q3 21 9 41t15 39H160Zm240-600h160v-80H400v80ZM720-40q-83 0-141.5-58.5T520-240q0-83 58.5-141.5T720-440q83 0 141.5 58.5T920-240q0 83-58.5 141.5T720-40Zm20-208v-112h-40v128l86 86 28-28-74-74Z'],
+        '0 -960 960 960'
+    ],
+    technologies: [
+        ['m240-198 79-32q-10-29-18.5-59T287-349l-47 32v119Zm160-42h160q18-40 29-97.5T600-455q0-99-33-187.5T480-779q-54 48-87 136.5T360-455q0 60 11 117.5t29 97.5Zm80-200q-33 0-56.5-23.5T400-520q0-33 23.5-56.5T480-600q33 0 56.5 23.5T560-520q0 33-23.5 56.5T480-440Zm240 242v-119l-47-32q-5 30-13.5 60T641-230l79 32ZM480-881q99 72 149.5 183T680-440l84 56q17 11 26.5 29t9.5 38v237l-199-80H359L160-80v-237q0-20 9.5-38t26.5-29l84-56q0-147 50.5-258T480-881Z'],
+        '0 -960 960 960'
+    ],
+    aboutMe: [
+        ['M80-160v-112q0-33 17-62t47-44q51-26 115-44t141-18q30 0 58.5 3t55.5 9l-70 70q-11-2-21.5-2H400q-71 0-127.5 17T180-306q-9 5-14.5 14t-5.5 20v32h250l80 80H80Zm542 16L484-282l56-56 82 82 202-202 56 56-258 258ZM400-480q-66 0-113-47t-47-113q0-66 47-113t113-47q66 0 113 47t47 113q0 66-47 113t-113 47Zm10 240Zm-10-320q33 0 56.5-23.5T480-640q0-33-23.5-56.5T400-720q-33 0-56.5 23.5T320-640q0 33 23.5 56.5T400-560Zm0-80Z'],
+        '0 -960 960 960'
+    ],
+    email: [
+        ['M638-80 468-250l56-56 114 114 226-226 56 56L638-80ZM480-520l320-200H160l320 200Zm0 80L160-640v400h206l80 80H160q-33 0-56.5-23.5T80-240v-480q0-33 23.5-56.5T160-800h640q33 0 56.5 23.5T880-720v174l-80 80v-174L480-440Zm0 0Zm0-80Zm0 80Z'],
+        '0 -960 960 960'
+    ],
+    gitHub: [
+        ['M 512 0 C 229.12 0 0 229.12 0 512 C 0 738.56 146.56 929.92 350.08 997.76 C 375.68 1002.24 385.28 986.88 385.28 973.44 C 385.28 961.28 384.64 920.96 384.64 878.08 C 256 901.76 222.72 846.72 212.48 817.92 C 206.72 803.2 181.76 757.76 160 745.6 C 142.08 736 116.48 712.32 159.36 711.68 C 199.68 711.04 228.48 748.8 238.08 764.16 C 284.16 841.6 357.76 819.84 387.2 806.4 C 391.68 773.12 405.12 750.72 419.84 737.92 C 305.92 725.12 186.88 680.96 186.88 485.12 C 186.88 429.44 206.72 383.36 239.36 347.52 C 234.24 334.72 216.32 282.24 244.48 211.84 C 244.48 211.84 287.36 198.4 385.28 264.32 C 426.24 252.8 469.76 247.04 513.28 247.04 C 556.8 247.04 600.32 252.8 641.28 264.32 C 739.2 197.76 782.08 211.84 782.08 211.84 C 810.24 282.24 792.32 334.72 787.2 347.52 C 819.84 383.36 839.68 428.8 839.68 485.12 C 839.68 681.6 720 725.12 606.08 737.92 C 624.64 753.92 640.64 784.64 640.64 832.64 C 640.64 901.12 640 956.16 640 973.44 C 640 986.88 649.6 1002.88 675.2 997.76 C 877.44 929.92 1024 737.92 1024 512 C 1024 229.12 794.88 0 512 0 Z'],
+        '0 0 1024 1024'
+    ],
+    linkedIn: [
+        ['M218.123 218.127h-37.931v-59.403c0-14.165-.253-32.4-19.728-32.4-19.756 0-22.779 15.434-22.779 31.369v60.43h-37.93V95.967h36.413v16.694h.51a39.907 39.907 0 0 1 35.928-19.733c38.445 0 45.533 25.288 45.533 58.186l-.016 67.013ZM56.955 79.27c-12.157.002-22.014-9.852-22.016-22.009-.002-12.157 9.851-22.014 22.008-22.016 12.157-.003 22.014 9.851 22.016 22.008A22.013 22.013 0 0 1 56.955 79.27m18.966 138.858H37.95V95.967h37.97v122.16ZM237.033.018H18.89C8.58-.098.125 8.161-.001 18.471v219.053c.122 10.315 8.576 18.582 18.89 18.474h218.144c10.336.128 18.823-8.139 18.966-18.474V18.454c-.147-10.33-8.635-18.588-18.966-18.453'],
+        '0 0 256 256'
+    ],
+    document: [
+        ['M560-80v-123l221-220q9-9 20-13t22-4q12 0 23 4.5t20 13.5l37 37q8 9 12.5 20t4.5 22q0 11-4 22.5T903-300L683-80H560Zm300-263-37-37 37 37ZM620-140h38l121-122-18-19-19-18-122 121v38ZM240-80q-33 0-56.5-23.5T160-160v-640q0-33 23.5-56.5T240-880h320l240 240v120h-80v-80H520v-200H240v640h240v80H240Zm280-400Zm241 199-19-18 37 37-18-19Z'],
+        '0 -960 960 960'
+    ],
+    check: [
+        ['M382-240 154-468l57-57 171 171 367-367 57 57-424 424Z'],
+        '0 -960 960 960'
+    ]
+}
